New and improved iteration I
made a split horizontal container, three different panel, put in the
buttons inside split container_panel_1 and the console control is inside
split_container_panel_2. I was able to make the Jlist and the buttons.
Right now nothing is functional but the basic outline is done, all we
have to do is apply the logic of each component is going to do.